Changed ExtendedKey class, not store any pointers
For some reason I decided the Extended Keys could be aware of their
own geneology via pointers to their parents... but that's unusable
if you're never guaranteeing that the parents will still be around.

Instead implemented it as a list of indices which is copied from
parent and then extended.