Man, extensible wallet crypto is complicated!
Here's my latest attempt at a unified way to apply encryption on:
   Master private encryption key with KDF (explicit IV)
   Private key using master private key (IV is addr)
   Private key using only KDF  (IV is addr)
   Public keys using a KDF  (explicit IV)
   P2SH scripts and Labels (parent chaincode is encryption key)

I think I finally found a way to do it: each of these requires four
8-byte identifiers:  KDF-algo-with-params hash, Encryption-algo hash,
encryption key source (sentinel or explicit), IV source (sentinel or
explicit).

KDF-algo-with-params will be stored in the wallet (after testing
system speed and finding good ones for user prefs), so will be a
list of supported encryption algorithms (which will be consistent
across all users).  Then 8 bytes is all you need.

To extend Armory to use different crypto or KDFs, you only need to
create such objects to be serialized into the file, and the hash of
their serialization will be their ID to use on the object.

This will probably have to be changed, but at least I have a coherent
first implementation of how to do this the way I envisioned